Refactoring the Restaurant Owners and Restaurant packages
