+++ src/main/java/br/com/techchallenge/adapters/controller/donoRestaurante/DonoRestauranteController.java
@@ -1,6 +1,6 @@
-package br.com.techchallenge.adapters;
+package br.com.techchallenge.adapters.controller.donoRestaurante;
 
-import br.com.techchallenge.adapters.response.DonoRestauranteResponse;
+import br.com.techchallenge.infra.dto.donoRestaurante.request.DonoRestauranteRequestDto;
 import br.com.techchallenge.application.BuscarDonoRestaurantePorIdUseCase;
 import br.com.techchallenge.application.CadastrarDonoRestauranteUseCase;
 import br.com.techchallenge.domain.DonoRestaurante;
@@ -10,7 +10,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDate;
-import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/donos-restaurante")
@@ -26,7 +25,7 @@ public class DonoRestauranteController {
     }
 
     @PostMapping
-    public ResponseEntity<String> cadastrar(@RequestBody DonoRestauranteRequest request) {
+    public ResponseEntity<String> cadastrar(@RequestBody DonoRestauranteRequestDto request) {
         try {
             DonoRestaurante dono = new DonoRestaurante(
                     request.nome(),
@@ -41,23 +40,6 @@ public class DonoRestauranteController {
             return new ResponseEntity<>("Dono de Restaurante cadastrado com sucesso", HttpStatus.CREATED);
         } catch (DataIntegrityViolationException e) {
             return new ResponseEntity<>("Dono de Restaurante já cadastrado com essas informações", HttpStatus.CONFLICT);
-        }
-    }
-
-    @GetMapping("/{id}")
-    public ResponseEntity<?> buscarPorId(@PathVariable Long id) {
-        Optional<DonoRestaurante> donoOptional = buscarDonoRestaurantePorIdUseCase.buscarPorId(id);
-
-        if (donoOptional.isPresent()) {
-            DonoRestaurante dono = donoOptional.get();
-            String message = String.format("Dono de Restaurante com o ID %d localizado com sucesso.", id);
-            DonoRestauranteResponse response = new DonoRestauranteResponse(
-                    message, dono.getNome(), dono.getEndereco(), dono.getEmail(), dono.getLogin()
-            );
-            return new ResponseEntity<>(response, HttpStatus.OK);
-        } else {
-            String mensagem = String.format("Dono de Restaurante com o ID %d não foi localizado.", id);
-            return new ResponseEntity<>(mensagem, HttpStatus.NOT_FOUND);
         }
     }
 }
