--- src/main/java/br/com/techchallenge/adapters/response/DonoRestauranteResponse.java
@@ -1,9 +1,0 @@
-package br.com.techchallenge.adapters.response;
-
-public record DonoRestauranteResponse(
-        String message,
-        String nome,
-        String endereco,
-        String email,
-        String login
-) {}
