+++ src/main/java/br/com/techchallenge/infra/dto/donoRestaurante/request/DonoRestauranteRequestDto.java
@@ -1,0 +1,9 @@
+package br.com.techchallenge.infra.dto.donoRestaurante.request;
+
+public record DonoRestauranteRequestDto(
+        String nome,
+        String endereco,
+        String email,
+        String login,
+        String senha
+) {}
