+++ src/main/java/br/com/techchallenge/infra/dto/donoRestaurante/response/DonoRestauranteResponseDto.java
@@ -1,0 +1,9 @@
+package br.com.techchallenge.infra.dto.donoRestaurante.response;
+
+public record DonoRestauranteResponseDto(
+        String nome,
+        String endereco,
+        String email,
+        String login,
+        String senha
+) {}
