+++ src/main/java/br/com/techchallenge/infra/dto/restaurante/request/RestauranteRequestDto.java
@@ -1,0 +1,7 @@
+package br.com.techchallenge.infra.dto.restaurante.request;
+
+public record RestauranteRequestDto(
+        String nome,
+        String endereco,
+        String tipoCozinha
+) {}
