+++ src/main/java/br/com/techchallenge/infra/dto/restaurante/response/RestauranteResponseDto.java
@@ -1,0 +1,7 @@
+package br.com.techchallenge.infra.dto.restaurante.response;
+
+public record RestauranteResponseDto(
+        String nome,
+        String endereco,
+        String tipoCozinha
+) {}
